GridCore: extract Column type to types file
