+++ packages/devextreme/js/__internal/grids/grid_core/editing/types.ts
@@ -1,4 +1,4 @@
-import type { Column } from '../columns_controller/m_columns_controller';
+import type { Column } from '../columns_controller/types';
 import type { Item, UserData } from '../data_controller/m_data_controller';
 import type { RowKey } from '../m_types';
 import type { INSERT_INDEX } from './const';
